Check when there is no active process

+++ sources/install/package_base.sh
@@ -252,7 +252,9 @@ function post_install() {
     rm -rfv /root/sources
     colorecho "Stop listening processes"
     LISTENING_PROCESSES=$(ss -lnpt | awk -F"," 'NR>1 {split($2,a,"="); print a[2]}')
-    kill -9 $LISTENING_PROCESSES
+    if [[ -n $LISTENING_PROCESSES ]]; then
+        kill -9 $LISTENING_PROCESSES
+    fi
     add-test-command "[[ $(sudo ss -lnpt | tail -n +2 | wc -l) -eq 0 ]]"
     colorecho "Sorting tools list"
     (head -n 1 /.exegol/installed_tools.csv && tail -n +2 /.exegol/installed_tools.csv | sort -f ) | tee /tmp/installed_tools.csv.sorted
